fix: change icon import
