+++ src/components/icons/index.ts
@@ -1,1 +1,1 @@
-export * from '@radix-ui/react-icons'
+export * from "@radix-ui/react-icons"
