+++ src/components/icons/social.tsx
@@ -1,0 +1,62 @@
+type IconProps = React.HTMLAttributes<SVGElement>
+
+export const GoogleIcon = (props: IconProps) => (
+  <svg
+    width="21"
+    height="21"
+    viewBox="0 0 20 21"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      d="M18.7509 10.6943C18.7509 9.97489 18.6913 9.44989 18.5624 8.90546H10.1794V12.1526H15.1C15.0009 12.9596 14.4652 14.1749 13.2747 14.9915L13.258 15.1002L15.9085 17.1125L16.0921 17.1305C17.7786 15.604 18.7509 13.3582 18.7509 10.6943Z"
+      fill="#4285F4"
+    />
+    <path
+      d="M10.1789 19.25C12.5896 19.25 14.6133 18.4722 16.0915 17.1305L13.2741 14.9916C12.5201 15.5069 11.5082 15.8666 10.1789 15.8666C7.81779 15.8666 5.81385 14.3402 5.0995 12.2305L4.99479 12.2392L2.23874 14.3295L2.2027 14.4277C3.67093 17.286 6.6868 19.25 10.1789 19.25Z"
+      fill="#34A853"
+    />
+    <path
+      d="M5.10002 12.2305C4.91153 11.6861 4.80245 11.1027 4.80245 10.5C4.80245 9.89716 4.91153 9.31385 5.0901 8.76941L5.08511 8.65346L2.29451 6.5296L2.20321 6.57216C1.59808 7.75829 1.25085 9.09026 1.25085 10.5C1.25085 11.9097 1.59808 13.2416 2.20321 14.4277L5.10002 12.2305Z"
+      fill="#FBBC05"
+    />
+    <path
+      d="M10.1789 5.13331C11.8555 5.13331 12.9864 5.84303 13.6313 6.43613L16.1511 4.025C14.6035 2.61528 12.5896 1.75 10.1789 1.75C6.68682 1.75 3.67094 3.71387 2.2027 6.57218L5.0896 8.76944C5.81387 6.65972 7.81782 5.13331 10.1789 5.13331Z"
+      fill="#EB4335"
+    />
+  </svg>
+)
+export const FacebookIcon = (props: IconProps) => (
+  <svg
+    width="21"
+    height="21"
+    viewBox="0 0 21 21"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <circle cx="10.5" cy="10.5" r="8.75" fill="#008CFF" />
+    <path
+      d="M13.7586 13.176L14.1472 10.7063H11.7158V9.10437C11.7158 8.42856 12.0548 7.76942 13.1438 7.76942H14.25V5.66687C14.25 5.66687 13.2466 5.5 12.2877 5.5C10.2843 5.5 8.97604 6.68309 8.97604 8.82403V10.7063H6.75V13.176H8.97604V19.1466C9.42294 19.215 9.88015 19.25 10.3459 19.25C10.8117 19.25 11.2689 19.215 11.7158 19.1466V13.176H13.7586Z"
+      fill="white"
+    />
+  </svg>
+)
+
+export const TwitterIcon = (props: IconProps) => (
+  <svg
+    width="21"
+    height="21"
+    viewBox="0 0 21 21"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      d="M15.0028 3.41666H17.4043L12.1575 9.4175L18.33 17.5833H13.497L9.71167 12.6308L5.38031 17.5833H2.97724L8.58924 11.1647L2.66797 3.41666H7.62366L11.0453 7.94346L15.0028 3.41666ZM14.1598 16.1448H15.4906L6.90056 4.77957H5.47252L14.1598 16.1448Z"
+      fill="#00200F"
+      fillOpacity="0.8"
+    />
+  </svg>
+)
